Create diff_viewer submodule
