+++ cascade/meta/diff_viewer/repo_diff_viewer.py
@@ -1,0 +1,172 @@
+import glob
+from typing import Dict, Tuple
+import os
+
+import pendulum
+from deepdiff import DeepDiff
+
+from ...base import MetaHandler, MetaFromFile
+from ..meta_viewer import MetaViewer
+from .base_diff_viewer import BaseDiffViewer
+
+
+class RepoDiffViewer(BaseDiffViewer):
+    def __init__(self, path) -> None:
+        """
+        Parameters
+        ----------
+        path: str
+            Path to the object which states to view
+        """
+        super().__init__(path)
+
+        self._default_depth = 2
+        self._default_diff_depth = 2
+
+    def _check_path(self, path: str, meta_type: Tuple) -> None:
+        if not os.path.isdir(path):
+            raise ValueError(f"Path `{path}` is not a directory")
+
+        # Check that meta of repo or line exists
+        # do not restrict extensions because meta handler would fail with
+        # informative message anyway
+        metas = glob.glob(os.path.join(path, "meta.*"))
+        if len(metas) == 0:
+            raise ValueError(f"There is no meta file in the directory provided: {path}")
+
+        # No idea how to handle multiple metas
+        # and what it means
+        if len(metas) > 1:
+            raise ValueError(f"Multiple meta files in the directory provided: {path}")
+
+        meta = MetaHandler.read(metas[0])
+        if "type" not in meta[0]:
+            raise ValueError(
+                f"Meta in file {metas[0]} has no `type` in its keys! "
+                "It may be that you are using DiffViewer on old "
+                "type of history logs before 0.10.0."
+            )
+
+        if not isinstance(meta, list):
+            raise ValueError(
+                f"Something is wrong with meta in {metas[0]} - it is not a list"
+            )
+
+        if "type" not in meta[0]:
+            raise ValueError(
+                f"Something is wrong with meta in {metas[0]} - no type key in it"
+            )
+
+        if meta[0]["type"] in meta_type:
+            raise ValueError("The folder you provided is neither the repo nor line")
+
+    def _read_objects(self, path: str) -> Dict[str, MetaFromFile]:
+        self._check_path(path, ("repo", "line"))
+
+        mev = MetaViewer(path, filt={"type": "model"})
+        objs = [meta for meta in mev]
+        objs = {meta[0]["name"]: meta for meta in objs}
+        return objs
+
+    def _layout(self):
+        def _table_row(name, prev_name):
+            date = self._objs[name][0]["created_at"]
+            date = pendulum.parse(date).format("DD MMMM YYYY HH:mm:ss")
+
+            obj = self._objs[name]
+            prev_obj = self._objs[prev_name]
+
+            diff = DeepDiff(prev_obj, obj).to_dict()
+
+            diff_str = ""
+            if "dictionary_item_added" in diff:
+                diff_str += f'+{len(diff["dictionary_item_added"])}'
+            if "values_changed" in diff:
+                diff_str += f'~{len(diff["values_changed"])}'
+            if "dictionary_item_removed" in diff:
+                diff_str += f'-{len(diff["dictionary_item_removed"])}'
+
+            return html.Tr(
+                children=[
+                    html.Th(date),
+                    html.Th(diff_str),
+                    html.Th(
+                        html.Details(
+                            children=[
+                                html.Summary(name),
+                                DashRenderjson(
+                                    id=f"{name}-data",
+                                    data={"": obj},
+                                    theme=self._json_theme,
+                                    max_depth=self._default_depth,
+                                ),
+                            ]
+                        )
+                    ),
+                ],
+                style={"text-align": "left"},
+            )
+
+        def line_table(line):
+            children = []
+            keys = list(lines[line].keys())
+            for name, prev_name in zip(keys, [keys[0], *keys[:-1]]):
+                children.append(_table_row(name, prev_name))
+
+            return html.Div(
+                children=[
+                    html.H3(line, style={"color": "#C92C6D"}),
+                    html.Table(id=f"table-{line}", children=children),
+                ],
+                style={"margin-bottom": "10px"},
+            )
+
+        try:
+            import dash
+        except ModuleNotFoundError:
+            self._raise_cannot_import_dash()
+        else:
+            from dash import dcc, html
+
+        try:
+            import dash_renderjson
+        except ModuleNotFoundError:
+            raise ModuleNotFoundError(
+                "Cannot import dash_renderjson. It is optional dependency for DiffViewer"
+                " and can be installed via `pip install dash_renderjson`"
+            )
+        else:
+            from dash_renderjson import DashRenderjson
+
+        self._objs = self._read_objects(self._path)
+
+        lines = dict()
+        for name in self._objs:
+            # /full/path/repo/line/model/file.ext
+            tail, _ = os.path.split(name)
+            tail, _ = os.path.split(tail)
+            _, line = os.path.split(tail)
+
+            if line in lines:
+                lines[line][name] = self._objs[name]
+            else:
+                lines[line] = {name: self._objs[name]}
+
+        return html.Div(
+            [
+                html.H1(
+                    children=f"DiffViewer in {self._path}",
+                    style={"textAlign": "center", **self._style},
+                ),
+                dcc.Dropdown(id="left-dropdown", options=list(self._objs.keys())),
+                dcc.Dropdown(id="rigth-dropdown", options=list(self._objs.keys())),
+                DashRenderjson(
+                    id="diff-json",
+                    data={"Nothing": "Nothing is selected!"},
+                    theme=self._json_theme,
+                    max_depth=self._default_diff_depth,
+                ),
+                *[line_table(line) for line in lines],
+            ],
+            style={"margin": "5%", **self._style},
+        )
