+++ cascade/meta/diff_viewer/workspace_diff_viewer.py
@@ -1,0 +1,11 @@
+from .base_diff_viewer import BaseDiffViewer
+
+
+class WorkspaceDiffViewer(BaseDiffViewer):
+    def __init__(self, path: str) -> None:
+        super().__init__(path)
+
+        self._default_depth = 2
+        self._default_diff_depth = 2
+
+    # def
